Experimental syntect 3 support.
SyntaxSet is now Send + Sync, allowing for a single instance to be
shared across all threads.  Since initialization was serialized and
per-thread previously, this is a significant win on systems with more
than a few cores, especially on smaller sites.

The regenerated syntax packdump is lacking Elm support due to a regexp
compile error.  Additional syntax support is also not yet hooked up.

This is part of #420's focus on improving performance.